requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
bcrypt==4.0.1
beautifulsoup4==4.13.4
certifi==2025.6.15
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
executing==2.2.0
faiss-cpu==1.11.0
fastapi==0.115.12
fastapi-cli==0.0.7
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.8.1
langchain==0.3.25
langchain-community==0.3.25
langchain-core==0.3.65
langchain-experimental==0.3.4
langchain-openai==0.3.24
langchain-tavily==0.2.4
langchain-text-splitters==0.3.8
langsmith==0.3.45
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.5.0
mypy==1.16.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.3.0
openai==1.88.0
orjson==3.10.18
packaging==24.2
pandas==2.3.0
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pypdf==5.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.7
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.46.2
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.9.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ujson==5.10.0
Unidecode==1.4.0
urllib3==2.5.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
yarl==1.20.1
zstandard==0.23.0
